Feature/authentication

+++ server/app.py
@@ -1,28 +1,50 @@
-from flask import request
+from flask import request, session
 from flask_restful import Resource
+
 from config import app, api
-from models import db, User, Exercise
+from models import User, Exercise, db
 from schemas import (
     user_schema,
     users_schema,
     exercise_schema,
-    exercises_schema,
+    exercises_schema
 )
 
+@app.before_request
+def check_if_logged_in():
 
-class Users(Resource):
+    open_endpoints = [
+        "signup",
+        "login",
+        "checksession",
+        "logout",
+        "static"
+    ]
 
-    # GET /users
-    def get(self):
-        users = User.query.all()
+    if request.endpoint in open_endpoints:
+        return
 
-        return users_schema.dump(users), 200
+    if "user_id" not in session:
+        return {
+            "error": "Unauthorized"
+        }, 401
 
+class Signup(Resource):
 
-    # POST /users
     def post(self):
 
         data = request.get_json()
+
+        # Check if username already exists
+        existing_user = User.query.filter_by(
+            username=data["username"]
+        ).first()
+
+        if existing_user:
+            return {
+                "error": "Username already exists"
+            }, 422
+
 
         user = User(
             username=data["username"],
@@ -30,13 +52,39 @@ class Users(Resource):
             age=data.get("age")
         )
 
+        # Hash password
+        user.set_password(data["password"])
+
         db.session.add(user)
         db.session.commit()
 
-        return user_schema.dump(user), 201
 
-    #Enpoint
+        # Log user in immediately
+        session["user_id"] = user.id
+
+
+        return user_schema.dump(user), 201
+    
+# Register signup route
+api.add_resource(Signup, "/signup")
+
+class Users(Resource):
+
+    # GET /users
+    def get(self):
+        user = User.query.get(session["user_id"])
+
+        if not user:
+            return {
+                "error": "User not found"
+            }, 404
+
+        return user_schema.dump(user), 200
+
+
+#Enpoint
 api.add_resource(Users, "/users")
+
 
 
 class UserByID(Resource):
@@ -45,6 +93,11 @@ class UserByID(Resource):
     def patch(self, id):
 
         user = User.query.get_or_404(id)
+
+        if user.id != session["user_id"]:
+            return {
+                "error": "Forbidden"
+            }, 403
 
         data = request.get_json()
 
@@ -64,6 +117,11 @@ class UserByID(Resource):
 
         user = User.query.get_or_404(id)
 
+        if user.id != session["user_id"]:
+            return {
+                "error": "Forbidden"
+            }, 403
+
         db.session.delete(user)
         db.session.commit()
 
@@ -73,14 +131,74 @@ class UserByID(Resource):
 
 api.add_resource(UserByID, "/users/<int:id>")
 
+class Login(Resource):
+
+    def post(self):
+
+        data = request.get_json()
+
+        user = User.query.filter_by(
+            username=data["username"]
+        ).first()
+
+        if user and user.check_password(data["password"]):
+
+            session["user_id"] = user.id
+
+            return user_schema.dump(user), 200
+
+        return {
+            "error": "Invalid username or password"
+        }, 401
+
+
+api.add_resource(Login, "/login")
+
+
+class CheckSession(Resource):
+
+    def get(self):
+
+        user_id = session.get("user_id")
+
+        if not user_id:
+            return {
+                "error": "Not logged in"
+            }, 401
+
+        user = User.query.get(user_id)
+
+        if not user:
+            return {
+                "error": "User not found"
+            }, 404
+
+        return user_schema.dump(user), 200
+
+
+api.add_resource(CheckSession, "/check_session")
+
+class Logout(Resource):
+
+    def delete(self):
+
+        session.pop("user_id", None)
+
+        return {
+            "message": "Logged out successfully"
+        }, 200
+
+
+api.add_resource(Logout, "/logout")
+
 
 class Exercises(Resource):
 
     # GET /exercises
     def get(self):
 
-        exercises = Exercise.query.all()
-
+        exercises = Exercise.query.filter_by(user_id=session["user_id"]).all()
+        
         return exercises_schema.dump(exercises), 200
 
 
@@ -94,7 +212,7 @@ class Exercises(Resource):
             category=data.get("category"),
             duration=data.get("duration"),
             calories_burned=data.get("calories_burned"),
-            user_id=data["user_id"]
+            user_id=session["user_id"]
         )
 
         db.session.add(exercise)
@@ -111,6 +229,11 @@ class ExerciseByID(Resource):
     def patch(self, id):
 
         exercise = Exercise.query.get_or_404(id)
+
+        if exercise.user_id != session["user_id"]:
+            return {
+                "error": "Forbidden"
+            }, 403
 
         data = request.get_json()
 
@@ -135,6 +258,10 @@ class ExerciseByID(Resource):
     def delete(self, id):
 
         exercise = Exercise.query.get_or_404(id)
+        if exercise.user_id != session["user_id"]:
+            return {
+                "error": "Forbidden"
+            }, 403
 
         db.session.delete(exercise)
         db.session.commit()
@@ -147,9 +274,6 @@ api.add_resource(
     ExerciseByID,
     "/exercises/<int:id>"
 )
-
-
-
